Django_tenant/requirement.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
certifi==2023.5.7
charset-normalizer==3.1.0
Django==3.2
django-sesame==3.1
django-tenant-schemas==1.11.0
django-tenants==3.4.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
ordered-set==4.1.0
pexpect==4.8.0
psycopg2-binary==2.9.6
ptyprocess==0.7.0
PyJWT==2.6.0
pytz==2023.3
requests==2.30.0
six==1.16.0
sqlparse==0.4.4
Strip==0.0.1.dev14
stripe==5.4.0
urllib3==2.0.2
